rename fathom-relative files
